Python/requirements.txt
python-dotenv==1.0.1
msal[broker]>=1.20,<2
gql[all]>=3.0.0,<4